# openapi.yaml
openapi: 3.0.3
info:
  title: Caretide Telemedicine API
  description: |
    REST API contract for the Caretide telemedicine portal frontend.
    Derived from `src/lib/api.ts` and `src/lib/types.ts`.

    All authenticated endpoints expect:
      `Authorization: Bearer <token>`

    Errors should respond with a JSON body:
      `{ "message": "human readable error" }`
  version: 1.0.0
  contact:
    name: Caretide
servers:
  - url: /api
    description: Same-origin (default for VITE_API_BASE_URL)
  - url: http://localhost:8080/api
    description: Local onsite backend
...
security:
  - bearerAuth: []
paths:
...
  /appointments:
    get:
      tags: [Appointments]
      summary: List appointments visible to the current user
      description: |
        Backend should filter by the caller's identity and role:
        - patient: only their own appointments
        - doctor:  only appointments where they are the provider
        - admin:   all appointments
      parameters:
        - in: query
          name: userId
          schema: { type: string }
          description: Optional — server should normally derive from auth token
        - in: query
          name: role
          schema: { $ref: "#/components/schemas/UserRole" }
      responses:
        "200":
          description: Appointments (sorted by scheduledAt asc)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Appointment" }
    post:
      tags: [Appointments]
      summary: Request a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [patientId, doctorId, scheduledAt, reason]
              properties:
                patientId: { type: string }
                patientName: { type: string }
                doctorId: { type: string }
                scheduledAt: { type: string, format: date-time }
                reason: { type: string }
      responses:
        "201":
          description: Appointment created (status=requested)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
  /appointments/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: string }
    patch:
      tags: [Appointments]
      summary: Update appointment status or notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { $ref: "#/components/schemas/AppointmentStatus" }
                notes: { type: string }
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Appointment" }
        "404": { $ref: "#/components/responses/Error" }

  # ---------------- RECORDS ----------------
...